updated get_Adder code

+++ owehub/owehub-main/services/apiUpdateAdderData.go
@@ -90,7 +90,7 @@ func HandleUpdateAdderDataRequest(resp http.ResponseWriter, req *http.Request) {
 		FormAndSendHttpResp(resp, "Invalid AdderCal Not Allowed", http.StatusBadRequest, nil)
 		return
 	}
-	if updateAdderDataReq.RepPercent <= 0 {
+	if updateAdderDataReq.RepPercent <= float64(0) {
 		err = fmt.Errorf("Invalid RepPercent Not Allowed")
 		log.FuncErrorTrace(0, "%v", err)
 		FormAndSendHttpResp(resp, "Invalid RepPercent Not Allowed", http.StatusBadRequest, nil)
